NVM of svm, native system contract virtual machine
