+++ core/state/statedb.go
@@ -230,3 +230,129 @@ func (s *Statedb) clearJournalAndRefund() {
 func (s *Statedb) GetCurrentLogs() []*types.Log {
 	return s.curLogs
 }
+
+// CreateAccount creates a new account in statedb.
+func (s *Statedb) CreateAccount(address common.Address) {
+	if object := s.getStateObject(address); object == nil {
+		object = newStateObject(address)
+		s.curJournal.append(createObjectChange{&address})
+		s.stateObjects[address] = object
+	}
+}
+
+// GetCodeHash returns the hash of the contract code associated with the specified address if any.
+// Otherwise, return an empty hash.
+func (s *Statedb) GetCodeHash(address common.Address) common.Hash {
+	if object := s.getStateObject(address); object != nil {
+		return common.BytesToHash(object.account.CodeHash)
+	}
+
+	return common.EmptyHash
+}
+
+// GetCode returns the contract code associated with the specified address if any.
+// Otherwise, return nil.
+func (s *Statedb) GetCode(address common.Address) []byte {
+	if object := s.getStateObject(address); object != nil {
+		return object.loadCode(s.trie)
+	}
+
+	return nil
+}
+
+// SetCode sets the contract code of the specified address if exists.
+func (s *Statedb) SetCode(address common.Address, code []byte) {
+	// EVM call SetCode after CreateAccount during contract creation.
+	// So, here the retrieved stateObj should not be nil.
+	if object := s.getStateObject(address); object != nil {
+		prevCode := object.loadCode(s.trie)
+		s.curJournal.append(codeChange{&address, prevCode})
+		object.setCode(code)
+	}
+}
+
+// GetCodeSize returns the size of the contract code associated with the specified address if any.
+// Otherwise, return 0.
+func (s *Statedb) GetCodeSize(address common.Address) int {
+	code := s.GetCode(address)
+	return len(code)
+}
+
+// GetState returns the value of the specified key in account storage if exists.
+// Otherwise, return empty hash.
+func (s *Statedb) GetState(address common.Address, key common.Hash) common.Hash {
+	value := s.GetData(address, key)
+	return common.BytesToHash(value)
+}
+
+// SetState adds or updates the specified key-value pair in account storage.
+func (s *Statedb) SetState(address common.Address, key common.Hash, value common.Hash) {
+	s.SetData(address, key, value.Bytes())
+}
+
+// Exist indicates whether the given account exists in statedb.
+// Note that it should also return true for suicided accounts.
+func (s *Statedb) Exist(address common.Address) bool {
+	return s.getStateObject(address) != nil
+}
+
+// Empty indicates whether the given account satisfies (balance = nonce = code = 0).
+func (s *Statedb) Empty(address common.Address) bool {
+	stateObj := s.getStateObject(address)
+	return stateObj == nil || stateObj.empty()
+}
+
+// Suicide marks the given account as suicided and clears the account balance.
+// Note the account's state object is still available until the state is committed.
+// Return true if the specified account exists, otherwise false.
+func (s *Statedb) Suicide(address common.Address) bool {
+	stateObj := s.getStateObject(address)
+	if stateObj == nil {
+		return false
+	}
+
+	s.curJournal.append(suicideChange{&address, stateObj.suicided, stateObj.getAmount()})
+
+	stateObj.setAmount(new(big.Int))
+	stateObj.suicided = true
+
+	return true
+}
+
+// HasSuicided returns true if the specified account exists and suicided, otherwise false.
+func (s *Statedb) HasSuicided(address common.Address) bool {
+	stateObj := s.getStateObject(address)
+	if stateObj == nil {
+		return false
+	}
+
+	return stateObj.suicided
+}
+
+// RevertToSnapshot reverts all state changes made since the given revision.
+func (s *Statedb) RevertToSnapshot(revid int) {
+	s.curJournal.revert(s, revid)
+}
+
+// Snapshot returns an identifier for the current revision of the statedb.
+func (s *Statedb) Snapshot() int {
+	return s.curJournal.snapshot()
+}
+
+// AddLog adds a log.
+func (s *Statedb) AddLog(log *types.Log) {
+	log.TxIndex = s.curTxIndex
+
+	s.curLogs = append(s.curLogs, log)
+}
+
+// AddRefund refunds the specified gas value
+func (s *Statedb) AddRefund(gas uint64) {
+	s.curJournal.append(refundChange{s.refund})
+	s.refund += gas
+}
+
+// GetRefund returns the current value of the refund counter.
+func (s *Statedb) GetRefund() uint64 {
+	return s.refund
+}
