+++ core/svm/evm/evm_test.go
@@ -85,7 +85,8 @@ func Test_SimpleStorage(t *testing.T) {
 	createContractTx, err := types.NewContractTransaction(address, new(big.Int), big.NewInt(1), 38, code)
 	assert.Equal(t, err, nil)
 
-	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, statedb, header, bcStore)}
+	stateDB := &StateDB{statedb}
+	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, stateDB, header, bcStore)}
 	receipt, err := svm.Process(createContractTx, 8)
 	// Validate receipt of contract creation.
 	assert.Equal(t, err, nil)
@@ -106,7 +107,8 @@ func Test_SimpleStorage(t *testing.T) {
 	callContractTx, err := types.NewMessageTransaction(address, contractAddr, new(big.Int), big.NewInt(1), 1, input)
 	assert.Equal(t, err, nil)
 
-	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, statedb, header, bcStore)}
+	stateDB.Statedb = statedb
+	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, stateDB, header, bcStore)}
 	receipt, err = svm.Process(callContractTx, 9)
 
 	// Validate receipt of contract call
@@ -123,7 +125,7 @@ func Test_SimpleStorage(t *testing.T) {
 	callContractTx, err = types.NewMessageTransaction(address, contractAddr, new(big.Int), big.NewInt(1), 1, input)
 	assert.Equal(t, err, nil)
 
-	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, statedb, header, bcStore)}
+	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, stateDB, header, bcStore)}
 	receipt, err = svm.Process(callContractTx, 10)
 
 	// Validate receipt contract call
@@ -140,7 +142,7 @@ func Test_SimpleStorage(t *testing.T) {
 	callContractTx, err = types.NewMessageTransaction(address, contractAddr, new(big.Int), big.NewInt(1), 1, input)
 	assert.Equal(t, err, nil)
 
-	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, statedb, header, bcStore)}
+	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, stateDB, header, bcStore)}
 	receipt, err = svm.Process(callContractTx, 11)
 
 	// Validate receipt of contract call
@@ -166,7 +168,8 @@ func Test_InsufficientBalance(t *testing.T) {
 	createContractTx, err := types.NewContractTransaction(address, new(big.Int), common.SeeleToFan, 38, code)
 	assert.Equal(t, err, nil)
 
-	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, statedb, header, bcStore)}
+	stateDB := &StateDB{statedb}
+	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, stateDB, header, bcStore)}
 	_, err = svm.Process(createContractTx, 8)
 	assert.Equal(t, err, vm.ErrInsufficientBalance)
 }
@@ -186,7 +189,8 @@ func Benchmark_CreateContract(b *testing.B) {
 	nonce := uint64(38)
 	createContractTx, _ := types.NewContractTransaction(address, amount, fee, nonce, code)
 
-	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, statedb, header, bcStore)}
+	stateDB := &StateDB{statedb}
+	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, stateDB, header, bcStore)}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		svm.Process(createContractTx, 8)
@@ -209,7 +213,8 @@ func Benchmark_CallContract(b *testing.B) {
 	nonce := uint64(38)
 	createContractTx, _ := types.NewContractTransaction(address, amount, fee, nonce, code)
 
-	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, statedb, header, bcStore)}
+	stateDB := &StateDB{statedb}
+	svm := &EVM{Evm: NewEVMByDefaultConfig(createContractTx, stateDB, header, bcStore)}
 	receipt, _ := svm.Process(createContractTx, 8)
 	contractAddr := common.BytesToAddress(receipt.ContractAddress)
 
@@ -219,7 +224,7 @@ func Benchmark_CallContract(b *testing.B) {
 	nonce = uint64(1)
 	callContractTx, _ := types.NewMessageTransaction(address, contractAddr, amount, fee, nonce, input)
 
-	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, statedb, header, bcStore)}
+	svm = &EVM{Evm: NewEVMByDefaultConfig(callContractTx, stateDB, header, bcStore)}
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		svm.Process(callContractTx, 9)
