+++ core/svm/svm.go
@@ -6,9 +6,11 @@
 package svm
 
 import (
+	"github.com/seeleteam/go-seele/contract/system"
 	"github.com/seeleteam/go-seele/core/state"
 	"github.com/seeleteam/go-seele/core/store"
 	"github.com/seeleteam/go-seele/core/svm/evm"
+	"github.com/seeleteam/go-seele/core/svm/native"
 	"github.com/seeleteam/go-seele/core/types"
 )
 
@@ -25,10 +27,15 @@ type Context struct {
 	BcStore     store.BlockchainStore
 }
 
-// NewSeeleVM implements a variety of vm, and you must ensure that the SVMTYPE is completed, otherwise the returns result is nil
+// NewSeeleVM implements a variety of vm, now only evm is completed
 func NewSeeleVM(ctx *Context) SeeleVM {
-	// TODO for other vm
-	return &evm.EVM{
-		Evm: evm.NewEVMByDefaultConfig(ctx.Tx, ctx.Statedb, ctx.BlockHeader, ctx.BcStore),
+	// NVM
+	_, ok := system.Contracts[ctx.Tx.Data.To]
+	if ok {
+		return native.NewNativeVM(ctx.Tx, ctx.Statedb, ctx.BlockHeader, ctx.BcStore)
 	}
+
+	// EVM
+	statedb := &evm.StateDB{Statedb: ctx.Statedb}
+	return &evm.EVM{Evm: evm.NewEVMByDefaultConfig(ctx.Tx, statedb, ctx.BlockHeader, ctx.BcStore)}
 }
